mejoras en el login

+++ api/users/users.controller.js
@@ -7,27 +7,36 @@ module.exports = {
   createUser,
   editPatch,
   deleteUser,
+  deleteSelfUser,
 };
 
 function getAllUsers(req, res) {
-  userModel
-    .find()
-    .then((response) => {
-      res.json(response);
-    })
-    .catch((err) => res.status(500).json(err));
+  if (req.currentUser.role === "admin") {
+    userModel
+      .find()
+      .then((response) => {
+        res.json(response);
+      })
+      .catch((err) => res.status(500).json(err));
+  } else {
+    res.status(403).send("No eres admin");
+  }
 }
 
 function getById(req, res) {
-  let userId = mongoose.Types.ObjectId.isValid(req.params.id);
+  if (req.currentUser.role === "admin") {
+    let userId = mongoose.Types.ObjectId.isValid(req.params.id);
 
-  if (userId) {
-    userModel
-      .findById(req.params.id)
-      .then((r) => res.json(r))
-      .catch((err) => res.status(500).json("Error en la base de datos"));
+    if (userId) {
+      userModel
+        .findById(req.params.id)
+        .then((r) => res.json(r))
+        .catch((err) => res.status(500).json("Error en la base de datos"));
+    } else {
+      res.status(404).send("Ningun usuario encontrado");
+    }
   } else {
-    res.status(404).send("Ningun usuario encontrado");
+    res.status(403).send("No eres admin");
   }
 }
 
@@ -58,24 +67,55 @@ function createUser(req, res) {
   }
 }
 function editPatch(req, res) {
-  let userId = mongoose.Types.ObjectId.isValid(req.params.id);
+  if (req.currentUser.role === "admin") {
+    let userId = mongoose.Types.ObjectId.isValid(req.params.id);
 
-  if (userId) {
-    userModel
-      .findByIdAndUpdate(req.params.id, req.body)
-      .then((r) => res.json(r))
-      .catch((err) => res.status(500).json("Error en la base de datos"));
+    if (userId) {
+      userModel
+        .findByIdAndUpdate(req.params.id, req.body)
+        .then((r) => res.json(r))
+        .catch((err) => res.status(500).json("Error en la base de datos"));
+    } else {
+      res.status(404).send("Ningun usuario encontrado");
+    }
   } else {
-    res.status(404).send("Ningun usuario encontrado");
+    res.status(403).send("No eres admin");
   }
 }
 
 function deleteUser(req, res) {
+  if (req.currentUser.role === "admin") {
+    let userId = mongoose.Types.ObjectId.isValid(req.params.id);
+    if (userId) {
+      userModel
+        .findByIdAndDelete(req.params.id)
+        .then((r) => res.json(r))
+        .catch((err) => res.status(500).json("Error en la base de datos"));
+    } else {
+      res.status(404).send("Ningun usuario encontrado");
+    }
+  } else {
+    res.status(403).send("No eres admin");
+  }
+}
+
+function deleteSelfUser(req, res) {
   let userId = mongoose.Types.ObjectId.isValid(req.params.id);
+
   if (userId) {
     userModel
-      .findByIdAndDelete(req.params.id)
-      .then((r) => res.json(r))
+      .findById(req.params.id)
+      .then((r) => {
+        if (r.email === req.currentUser.email) {
+          userModel
+            .findByIdAndDelete(req.params.id)
+            .then((r) => res.send("Eliminado con exito"));
+        } else {
+          res
+            .status(403)
+            .send("El email actual no coincide con el email a eliminar");
+        }
+      })
       .catch((err) => res.status(500).json("Error en la base de datos"));
   } else {
     res.status(404).send("Ningun usuario encontrado");
